requirements.txt
python-can>=4.4.0
pyserial>=3.5
supabase>=2.0.0
python-dotenv>=1.0.0
